Fix lint issues and make API prefixing consistent.
Put the API namespace directly in the named calls, which keeps it
consistent with all the other API definitions. API namespaces are also
unlikely to change often (if ever), so hardcoding them in the API will
force the developer to think about whether they should be changing the
prefix instead of haphazardly updating a single variable.
Also updates some linting issue relating to object definition positions
(opening brace on newline) on the affected lines.